Move `SharesReader` to the `io.spine.examples.shareaware.share` package.
