+++ server/src/main/java/io/spine/examples/shareaware/server/market/MarketData.java
@@ -28,7 +28,7 @@ package io.spine.examples.shareaware.server.market;
 
 import com.google.common.collect.ImmutableList;
 import io.spine.examples.shareaware.share.Share;
-import io.spine.examples.shareaware.sharesreader.SharesReader;
+import io.spine.examples.shareaware.share.SharesReader;
 import io.spine.money.Money;
 
 import java.io.File;
